+++ src/Functionality.Ioc.Autofac.Test/TypeAndFactoryListTest.cs
@@ -1,5 +1,4 @@
 ﻿using Autofac;
-using Autofac.Core;
 using AutoFixture;
 using AutoFixture.AutoMoq;
 using NUnit.Framework;
@@ -7,7 +6,7 @@ using Phoenix.Functionality.Ioc.Autofac;
 
 namespace Functionality.Ioc.Autofac.Test;
 
-public class TypeListSourceTest
+public class TypeAndFactoryListTest
 {
 	#region Setup
 
@@ -57,49 +56,74 @@ public class TypeListSourceTest
 	#endregion
 
 	#region Tests
-
+	
 	[Test]
-	public void GetRegisteredTypes()
+	public void GetRegisteredTypesAndFactories()
 	{
 		// Arrange
 		var builder = new ContainerBuilder();
-		builder.RegisterSource<TypeListSource<ISomething>>();
-		builder.RegisterType<Anything_01>().As<ISomething>().SingleInstance();
-		builder.RegisterType<Anything_02>().As<ISomething>().SingleInstance();
-		builder.RegisterType<Anything_03>().As<ISomething>().SingleInstance();
+		builder.RegisterSource<TypeAndFactoryListSource<ISomething>>();
+		builder.RegisterType<Anything_01>().AsSelf().As<ISomething>().SingleInstance();
+		builder.RegisterType<Anything_02>().AsSelf().As<ISomething>().SingleInstance();
+		builder.RegisterType<Anything_03>().AsSelf().As<ISomething>().SingleInstance();
 		builder.RegisterType<NotPart>().SingleInstance();
 		var container = builder.Build();
 
 		// Act
-		var types = container.Resolve<TypeList<ISomething>>();
+		var typesAndFactories = container.Resolve<TypeAndFactoryList<ISomething>>();
 		
 		// Assert
 		Assert.Multiple
 		(
 			() =>
 			{
-				Assert.NotNull(types);
-				Assert.IsNotEmpty(types);
-				Assert.That(types, Has.Count.EqualTo(3));
+				Assert.NotNull(typesAndFactories);
+				Assert.IsNotEmpty(typesAndFactories);
+				Assert.That(typesAndFactories, Has.Count.EqualTo(3));
+
+				var types = typesAndFactories.Select(tuple => tuple.Type).ToArray();
 				Assert.That(types, Has.Member(typeof(Anything_01)));
 				Assert.That(types, Has.Member(typeof(Anything_02)));
 				Assert.That(types, Has.Member(typeof(Anything_03)));
 				Assert.That(types, Has.No.Member(typeof(NotPart)));
+
+				foreach (var (type, factory) in typesAndFactories)
+				{
+					var instance = factory.Invoke();
+					Assert.That(instance, Is.TypeOf(type));
+				}
 			}
 		);
 	}
-	
+
+	[Test]
+	public void NotRegisteringTypeAsSelfThrows()
+	{
+		// Arrange
+		var builder = new ContainerBuilder();
+		builder.RegisterSource<TypeAndFactoryListSource<ISomething>>();
+		builder.RegisterType<Anything_01>()/*.AsSelf()*/.As<ISomething>().SingleInstance(); //! Don't register AsSelf().
+		var container = builder.Build();
+
+		// Act
+		var (_, factory) = container.Resolve<TypeAndFactoryList<ISomething>>().Single();
+		var exception = Assert.Catch<global::Autofac.Core.Registration.ComponentNotRegisteredException>(() => factory.Invoke());
+		
+		// Assert
+		Assert.That(exception, Is.Not.Null);
+	}
+
 	[Test]
 	public void ObtainingTypeDoesNotCreateInstance()
 	{
 		// Arrange
 		var builder = new ContainerBuilder();
-		builder.RegisterSource<TypeListSource<ISomething>>();
-		builder.RegisterType<Anything>().As<ISomething>().SingleInstance();
+		builder.RegisterSource<TypeAndFactoryListSource<ISomething>>();
+		builder.RegisterType<Anything>().AsSelf().As<ISomething>().SingleInstance();
 		var container = builder.Build();
 
 		// Act
-		var types = container.Resolve<TypeList<ISomething>>();
+		var types = container.Resolve<TypeAndFactoryList<ISomething>>();
 		
 		// Assert
 		Assert.NotNull(types);