+++ src/Functionality.Ioc.Autofac/⬙/CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
-## 2.2.1
+## 2.3.0
 
-:calendar: _2023-09-??_
+:calendar: _2023-09-09_
+
+### Added
+
+- New `TypeAndFactoryList{T}` that can be used to obtain the types of registered services (as opposed to their instances) alongside a factory for actually creating instances of the type via the appropriate **IRegistrationSource** `TypeAndFactoryListSource{T}`.
 
 ### Fixed